remove entities and methods files
