+++ stellar/.gitignore
@@ -1,5 +1,5 @@
 .DS_Store
 stellar
 yol.db
-seed.hex
+platformseed.hex
 sbc.db
