+++ stellar/database/contractentities.go
@@ -9,6 +9,62 @@ import (
 	"github.com/boltdb/bolt"
 )
 
+// the contractor super struct comprises of various entities within it. Its a
+// super class because combining them results in less duplication of code
+// TODO: in some ways, the Name, LoginUserName and LoginPassword fields can be
+// devolved into a separate User struct, that would result in less duplication as
+// well
+type ContractEntity struct {
+	// User defines common params such as name, seed, publickey
+	U User
+	// the name of the contractor / company that is contracting
+	// A contractor is party who proposes a specific some of money towards a
+	// particular project. This is the actual amount that the investors invest in.
+	// This ideally must include the developer fee within it, so that investors
+	// don't have to invest in two things. It would also make sense because the contractors
+	// sometimes would hire developers themselves.
+	Contractor bool
+	// A guarantor is somebody who can assure investors that the school will get paid
+	// on time. This authority should be trusted and either should be vetted by the law
+	// or have a multisig paying out to the investors beyond a certain timeline if they
+	// don't get paid by the school. This way, the guarantor can be anonymous, like the
+	// nice Pineapple Fund guy. THis can also be an insurance company, who is willing to
+	// guarantee for specific school and the school can pay him out of chain / have
+	// that as fee within the contract the originator
+	Developer bool
+	// A developer is someone who installs the required equipment (Raspberry Pi,
+	// network adapters, anti tamper installations and similar) In the initial
+	// orders, this will be us, since we'd be installign the pi ourselves, but in
+	// the future, we expect third party developers / companies to do this for us
+	// and act in a decentralized fashion. This money can either be paid out of chain
+	// in fiat or can be a portion of the funds the investors chooses to invest in.
+	// a contractor may also employ developers by himself, so this entity is not
+	// strictly necessary.
+	Originator bool
+	// An Originator is an entity that will start a project and get a fixed fee for
+	// rendering its service. An Originator's role is not restricted, the originator
+	// can also be the developer, contractor or guarantor. The originator should take
+	// the responsibility of auditing the requirements of the project - panel size,
+	// location, number of panels needed, etc. He then should ideally be able to fill
+	// out some kind of form on the website so that the originator's proposal is live
+	// and shown to potential investors. The originators get paid only when the order
+	// is live, else they can just spam, without any actual investment
+	Guarantor     bool
+	PastContracts []Contract
+	// list of all the contracts that the contractor has won in the past
+	ProposedContracts []Contract
+	// the Originator proposes a contract which will then be taken up
+	// by a contractor, who publishes his own copy of the proposed contract
+	// which will be the set of contracts that will be sent to auction
+	PresentContracts []Contract
+	// list of all contracts that the contractor is presently undertaking1
+	PastFeedback []Feedback
+	// feedback received on the contractor from parites involved in the past
+	// What kind of proof do we want from the company? KYC?
+	// maybe we could have a photo op like exchanges do these days, with the owner
+	// holding up his drivers' license or similar
+}
+
 func newContractEntityHelper(uname string, pwd string, Name string, Address string, Description string, role string) (ContractEntity, error) {
 	// call this after the user has failled in username and password. Store hashed password
 	// in the database
