+++ stellar/database/contracts.go
@@ -1,0 +1,5 @@
+package database
+
+type Contract struct {
+  O Order
+}
