+++ stellar/database/orders.go
@@ -15,6 +15,41 @@ import (
 	"github.com/boltdb/bolt"
 )
 
+// Order is what is advertised on the frontend where investors can choose what
+// projects to invest in.
+// the idea of an Order should be constructed by ourselves based on some preset
+// parameters. It should hold a contractor, developers, guarantor, originator
+// along with the fields already defined. The other fields should be displayed
+// on the frontend.
+// Order is the structure that is actually stored in the database
+// Order is a subset of the larger struct Contractor, which is still a TODO
+type Order struct {
+	Index         int  // an Index to keep quick track of how many orders exist
+	PanelSize     string  // size of the given panel, for diplsaying to the user who wants to bid stuff
+	TotalValue    int     // the total money that we need from investors
+	Location      string  // where this specific solar panel is located
+	MoneyRaised   int     // total money that has been raised until now
+	Years         int     // number of years the recipient has chosen to opt for
+	Metadata      string  // any other metadata can be stored here
+	Live          bool    // check to see whether the current order is live or not
+	Origin        bool    // if this order is an originated order
+	Votes         int     // the number of votes of a proposed contract
+	PaidOff       bool    // whether the asset has been paidoff by the recipient
+	INVAssetCode  string  // once all funds have been raised, we need to set assetCodes
+	DEBAssetCode  string  // once all funds have been raised, we need to set assetCodes
+	PBAssetCode   string  // once all funds have been raised, we need to set assetCodes
+	BalLeft       float64 // denotes the balance left to pay by the party
+	DateInitiated string  // date the order was created
+	DateFunded    string  // date when the order was funded
+	DateLastPaid  string  // date the order was last paid
+	// instead of just holding the recipient's name here, we can hold the recipient
+	OrderRecipient Recipient
+	// also have an array of investors to keep track of who invested in these projects
+	OrderInvestors []Investor
+	// TODO: have an investor and recipient relation here
+	// Percentage raised is not stored in the database since that can be calculated by the UI
+}
+
 // NewOrder creates a new order struct with the order parameters pased to the function
 // quite ugly with all the parameters passed, would be nice if we could rewrite this
 // in a nicer way
