+++ stellar/database/users.go
@@ -10,6 +10,41 @@ import (
 	xlm "github.com/YaleOpenLab/smartPropertyMVP/stellar/xlm"
 )
 
+// the user structure  houses all entities that are of type "User". This contains
+// commonly used functions so that we need not repeat the ssame thing for every instance.
+type User struct {
+	Index int
+	// default index, gets us easy stats on how many people are there and stuff,
+	// don't want to omit this
+	Seed string
+	// Seed is the equivalent of a private key in stellar (stellar doesn't expose private keys)
+	// do we make seed optional like that for the Recipient? Couple things to consider
+	// here: if the recipient loses the publickey, it can nver send DEBTokens back
+	// to the issuer, so it would be as if it reneged on the deal. Do we count on
+	// technically less sound people to hold their public keys safely? I suggest
+	// this would be  difficult in practice, so maybe enforce that they need to hold|
+	// their account on the platform?
+	Name string
+	// Name of the primary stakeholder involved (principal trustee of school, for eg.)
+	PublicKey string
+	// PublicKey denotes the public key of the recipient
+	LoginUserName string
+	// the username you use to login to the platform
+	LoginPassword string
+	// the password, which you use to authenticate on the platform
+	Address string
+	// the registered address of the above company
+	Description string
+	// Does the contractor need to have a seed and a publickey?
+	// we assume that it does in this case and proceed.
+	// information on company credentials, their experience
+	Image string
+	// image can be company logo, founder selfie
+	// hash of the password in reality
+	FirstSignedUp string
+	// auto generated timestamp
+}
+
 // User is a metastrucutre that contains commonyl used keys within a single umbrella
 // so that we can import it wherever needed.
 func NewUser(uname string, pwd string, Name string) (User, error) {
